Refine ECR repository IAM policy
Allow reader accounts access to the following actions:

*   ecr:DescribeImages
*   ecr:DescribeRepositories
*   ecr:ListImages

Also update README.md, variables.tf, and versions.tf.

This change will allow users on the development and test accounts to
view ECR repos and images that reside on the production account without
having production credentials.

I do not believe this capability is available through the AWS console,
but it is available through the AWS CLI and API.